+++ vm/src/cairo_run.rs
@@ -1,6 +1,9 @@
 use crate::{
     hint_processor::hint_processor_definition::HintProcessor,
-    types::{builtin_name::BuiltinName, layout_name::LayoutName, program::Program},
+    types::{
+        builtin_name::BuiltinName, layout::CairoLayoutParams, layout_name::LayoutName,
+        program::Program,
+    },
     vm::{
         errors::{
             cairo_run_errors::CairoRunError, runner_errors::RunnerError, vm_exception::VmException,
@@ -26,6 +29,9 @@ pub struct CairoRunConfig<'a> {
     pub trace_enabled: bool,
     pub relocate_mem: bool,
     pub layout: LayoutName,
+    /// The `dynamic_layout_params` argument should only be used with dynamic layout.
+    /// It is ignored otherwise.
+    pub dynamic_layout_params: Option<CairoLayoutParams>,
     pub proof_mode: bool,
     pub secure_run: Option<bool>,
     pub disable_trace_padding: bool,
@@ -43,6 +49,7 @@ impl<'a> Default for CairoRunConfig<'a> {
             secure_run: None,
             disable_trace_padding: false,
             allow_missing_builtins: None,
+            dynamic_layout_params: None,
         }
     }
 }
@@ -65,6 +72,7 @@ pub fn cairo_run_program_with_initial_scope(
     let mut cairo_runner = CairoRunner::new(
         program,
         cairo_run_config.layout,
+        cairo_run_config.dynamic_layout_params.clone(),
         cairo_run_config.proof_mode,
         cairo_run_config.trace_enabled,
     )?;
@@ -151,6 +159,7 @@ pub fn cairo_run_pie(
     let mut cairo_runner = CairoRunner::new(
         &program,
         cairo_run_config.layout,
+        cairo_run_config.dynamic_layout_params.clone(),
         false,
         cairo_run_config.trace_enabled,
     )?;
@@ -223,6 +232,7 @@ pub fn cairo_run_fuzzed_program(
     let mut cairo_runner = CairoRunner::new(
         &program,
         cairo_run_config.layout,
+        cairo_run_config.dynamic_layout_params.clone(),
         cairo_run_config.proof_mode,
         cairo_run_config.trace_enabled,
     )?;